ChatBot/requirements.txt

flask
numpy
pandas
pyimgur
gunicorn
requests
matplotlib
mplfinance
line-bot-sdk